requirements.txt
Flask-JWT-Extended==4.2.3
Flask-JWT==0.3.2
Flask-Migrate==3.1.0
Flask-OAuthlib==0.9.6
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask==2.0.1
SQLAlchemy==1.4.22
Werkzeug==2.0.1
alembic==1.6.5
flasgger
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.26.1
marshmallow==3.13.0
psycopg2
python-dotenv
passlib==1.7.4

